requirements.txt
pika
requests
newspaper3k
redis
nltk
scikit-learn
pymongo
dateutil
tensorflow
json-rpc
